pyproject.toml
[project]
name = "librarian"
version = "0.1.0"
description = "Personal agent design reference — compiled wiki, MCP server, Claude Code skills"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "fastmcp>=0.4.0",
    "duckdb>=1.0.0",
    "pdfplumber>=0.11.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.27.0",
    "structlog>=24.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "beautifulsoup4>=4.12.0",
    "python-frontmatter>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "playwright>=1.44.0",
]
viz = [
    "streamlit>=1.35.0",
    "st-cytoscape>=0.0.3",
    "networkx>=3.3",
    "python-frontmatter>=1.1.0",
]
api = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "sentence-transformers>=3.0.0",
    "umap-learn>=0.5.6",
    "watchfiles>=0.21.0",
    "websockets>=12.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
    "google-genai>=1.0.0",
    "ollama>=0.4.0",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.4.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-watch>=4.2.0",
    "httpx>=0.27.0",
    "playwright>=1.44.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "ANN", "B", "SIM"]
ignore = ["ANN101", "ANN102"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
